moving some files around in prep for building replay
