+++ app/CMakeLists.txt
@@ -9,6 +9,11 @@ option(WMTK_APP_SEC "surface shortest edge collapse" ON)
 option(WMTK_APP_QSLIM "surface qslim simplification" ON)
 option(WMTK_APP_UNIT_TESTS "unit tests for applications" ON)
 option(WMTK_APP_INTERIOR_TET_OPT "interior tet opt application" ON)
+option(WMTK_APP_REPLAYER "replaying existing mesh deformation application" ON)
+
+if(WMTK_APP_REPLAYER)
+add_subdirectory(replayer)
+endif()
 
 if(WMTK_APP_INTERIOR_TET_OPT)
 add_subdirectory(interior_tet_opt)
