+++ src/wmtk/AttributeCollectionRecorder.cpp
