+++ src/wmtk/utils/Hdf5Utils.cpp
@@ -1,4 +1,4 @@
-#include <wmtk/utils/Hdf5Utils.h>
+#include <wmtk/utils/Hdf5Utils.hpp>
 #include <wmtk/utils/Logger.hpp>
 namespace wmtk {
 // checks whether the file holding the dataset exists
@@ -7,25 +7,20 @@ bool does_dataset_exist(const HighFive::File& file, const std::string& name)
     auto obj_names = file.listObjectNames();
     if (file.exist(name)) {
         return true;
-        ;
     }
-    // for whaterver reason file.exist doesn't work with my invocation so doing it the dumb way
+    // MTAO: if, for whaterver reason, file.exist doesn't work with my
+    // invocation so doing it manually
     for (auto&& n : obj_names) {
         if (n == name) {
             if (HighFive::ObjectType::Dataset == file.getObjectType(name)) {
                 return true;
-            } else {
-                logger().error(
-                    "create_dataset: {} had root node {} but it was not a dataset",
-                    file.getName(),
-                    name);
             }
         }
     }
     return false;
 }
 HighFive::DataSet
-create_dataset(HighFive::File& file, const std::string& name, const HighFive::DataType& datatype)
+create_extendable_dataset(HighFive::File& file, const std::string& name, const HighFive::DataType& datatype)
 {
     if (does_dataset_exist(file, name)) {
         auto ds = file.getDataSet(name);
